Put together various pieces of the existing modules (#18)
This is a large overhaul of modules, putting together various pieces.
Currently, the modules are built as object files, so does the athena module. 
In the future, the athena module should be built as a library.
The current PR compiles but has not been tested. 
Testing will be performed in future PRs.